requirements.txt
music21==9.1.0
mido==1.3.2
Pillow==10.3.0